Remove previous logger and implement new file logger
