+++ src/AudioPluginProcessor.h
@@ -2,10 +2,28 @@
 
 #include "juce_audio_processors/juce_audio_processors.h"
 
+/**
+ * A global function to push log message to the logger.
+ * In a production build, this function would do nothing. No logs are recorded.
+ * Must call this function instead of invoking `juce::Logger::writeToLog()`
+ * to reduce unnecessary operations in a release build.
+ * @param msg log message to log.
+ */
+inline static void dlog(juce::StringRef msg) {
+#if DEBUG
+  juce::String s{};
+  s << '[' << juce::Time::getCurrentTime().toString(true, true, true, true)
+    << "] " << msg;
+  juce::Logger::writeToLog(s);
+#else
+  juce::ignoreUnused(msg);
+#endif
+}
+
 class AudioPluginProcessor : public juce::AudioProcessor {
  public:
   AudioPluginProcessor();
-  ~AudioPluginProcessor() override = default;
+  ~AudioPluginProcessor() override;
 
   void prepareToPlay(double sampleRate, int samplesPerBlock) override;
   void releaseResources() override;
@@ -44,5 +62,15 @@ class AudioPluginProcessor : public juce::AudioProcessor {
   void setStateInformation(const void* data, int sizeInBytes) override;
 
  private:
+  std::unique_ptr<juce::FileLogger> logger {
+#if DEBUG
+    juce::FileLogger::createDateStampedLogger(
+        "SoftVelvet", "runtime-log", ".log",
+        "New Instance of SoftVelvet Audio Plugin Initialised")
+#else
+    nullptr
+#endif
+  };
+
   JUCE_DECLARE_NON_COPYABLE_WITH_LEAK_DETECTOR(AudioPluginProcessor)
 };
